+++ temporian/implementation/numpy/operators/unary.py
@@ -3,7 +3,7 @@ from typing import Dict
 
 import numpy as np
 
-from temporian.core.operators.unary import (
+from temporian.core.operators.unary.unary import (
     BaseUnaryOperator,
     AbsOperator,
     InvertOperator,
